requirements.txt
django
psycopg2
Pillow
stripe
python-dotenv
flake8
djangorestframework